refactor(config): rely on tombi for toml editing
