+++ crates/project-model/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod macro_def;
 pub mod project_manifest;
-mod toml_manifest;
 mod toml_workspace;
 
 use std::collections::VecDeque;
@@ -12,12 +11,6 @@ use base_db::{
 };
 use itertools::Itertools;
 use rustc_hash::{FxHashMap, FxHashSet};
-pub use toml_manifest::{
-    TomlManifestDiagnostic, TomlManifestField, TomlManifestFieldCompletionContext,
-    TomlManifestPath, toml_manifest_diagnostics, toml_manifest_field_at_offset,
-    toml_manifest_field_completion_context, toml_manifest_fields, toml_manifest_path_at_offset,
-    toml_manifest_paths,
-};
 use triomphe::Arc;
 use utils::paths::{AbsPathBuf, sort_and_remove_subfolders};
 use vfs::{FileSetConfig, VfsPath};
