--- crates/project-model/src/toml_manifest/mod.rs
@@ -1,10 +1,0 @@
-mod schema;
-mod spans;
-
-pub(crate) use schema::TomlManifestSchema;
-pub use schema::{TomlManifestDiagnostic, toml_manifest_diagnostics};
-pub use spans::{
-    TomlManifestField, TomlManifestFieldCompletionContext, TomlManifestPath,
-    toml_manifest_field_at_offset, toml_manifest_field_completion_context, toml_manifest_fields,
-    toml_manifest_path_at_offset, toml_manifest_paths,
-};
