+++ src/global_state/main_loop.rs
@@ -233,7 +233,6 @@ impl GlobalState {
             )
             .on_latency_sensitive::<SemanticTokensRangeRequest>(handle_semantic_tokens_range)
             .on::<DocumentSymbolRequest>(handle_document_symbol)
-            .on::<DocumentLinkRequest>(handle_document_link)
             .on::<FoldingRangeRequest>(handle_folding_ranges)
             .on::<DocumentDiagnosticRequest>(handle_document_diagnostic)
             .on::<WorkspaceDiagnosticRequest>(handle_workspace_diagnostic)
