+++ src/tests.rs
@@ -10,9 +10,9 @@ use lsp_types::{
     CodeActionOrCommand, CodeActionParams, DiagnosticClientCapabilities,
     DidChangeTextDocumentParams, DidOpenTextDocumentParams, DidSaveTextDocumentParams,
     DocumentDiagnosticParams, DocumentDiagnosticReport, DocumentDiagnosticReportResult,
-    DocumentLinkParams, DocumentSymbolParams, DocumentSymbolResponse, FoldingRange,
-    FoldingRangeParams, GotoDefinitionParams, GotoDefinitionResponse, Hover, HoverParams, Position,
-    ProgressParams, PublishDiagnosticsParams, Range, SemanticTokensParams, SemanticTokensResult,
+    DocumentSymbolParams, DocumentSymbolResponse, FoldingRange, FoldingRangeParams,
+    GotoDefinitionParams, GotoDefinitionResponse, Hover, HoverParams, Position, ProgressParams,
+    PublishDiagnosticsParams, Range, SemanticTokensParams, SemanticTokensResult,
     TextDocumentClientCapabilities, TextDocumentContentChangeEvent, TextDocumentIdentifier,
     TextDocumentItem, TextDocumentPositionParams, Url, VersionedTextDocumentIdentifier,
     WorkDoneProgressParams, WorkspaceClientCapabilities, WorkspaceDiagnosticParams,
@@ -21,9 +21,9 @@ use lsp_types::{
         DidChangeTextDocument, DidOpenTextDocument, DidSaveTextDocument, Exit, Notification as _,
     },
     request::{
-        CodeActionRequest, Completion, DocumentDiagnosticRequest, DocumentLinkRequest,
-        DocumentSymbolRequest, FoldingRangeRequest, GotoDefinition, HoverRequest, Request as _,
-        SemanticTokensFullRequest, Shutdown, WorkspaceDiagnosticRequest,
+        CodeActionRequest, DocumentDiagnosticRequest, DocumentSymbolRequest, FoldingRangeRequest,
+        GotoDefinition, HoverRequest, Request as _, SemanticTokensFullRequest, Shutdown,
+        WorkspaceDiagnosticRequest,
     },
 };
 use serde::de::DeserializeOwned;
@@ -1586,7 +1586,7 @@ fn project_manifest_is_not_diagnosed_as_systemverilog() {
     };
     let temp_dir = TempDir::new("manifest-diagnostics");
     let manifest_text = "top_modules = [\"top\"]\nsources = [\"rtl\"]\n";
-    let manifest_path = temp_dir.path().join("vizsla.toml");
+    let manifest_path = temp_dir.path().join("vizsla_config.toml");
     fs::write(&manifest_path, manifest_text).unwrap();
     fs::create_dir_all(temp_dir.path().join("rtl")).unwrap();
 
@@ -1678,681 +1678,6 @@ fn project_manifest_is_not_diagnosed_as_systemverilog() {
 
     let (_, diagnostics) = recv_document_diagnostics(&client, request_id);
     assert!(diagnostics.is_empty(), "manifest must not receive slang diagnostics: {diagnostics:?}");
-
-    shutdown_test_server(&client, server_thread);
-}
-
-#[test]
-fn project_manifest_reports_toml_schema_diagnostics() {
-    let pull_caps = ClientCapabilities {
-        text_document: Some(TextDocumentClientCapabilities {
-            diagnostic: Some(DiagnosticClientCapabilities::default()),
-            ..Default::default()
-        }),
-        ..Default::default()
-    };
-    let temp_dir = TempDir::new("manifest-schema-diagnostics");
-    let manifest_text = "source = [\"rtl\"]\n";
-    let manifest_path = temp_dir.path().join("vizsla.toml");
-    fs::write(&manifest_path, manifest_text).unwrap();
-
-    let root_path = temp_dir.path().to_path_buf();
-    let opt = Opt {
-        process_name: "vizsla-test".to_string(),
-        log: "error".to_string(),
-        log_filename: None,
-    };
-    let config = config::Config::new(
-        opt,
-        root_path.clone(),
-        pull_caps,
-        vec![root_path],
-        UserConfig::default(),
-        Vec::new(),
-    );
-
-    let (server, client) = Connection::memory();
-    let server_thread = thread::spawn(move || main_loop::main_loop(config, server));
-    let manifest_uri = to_proto::url_from_abs_path(manifest_path.as_path()).unwrap();
-
-    client
-        .sender
-        .send(Message::Notification(Notification::new(
-            DidOpenTextDocument::METHOD.to_string(),
-            DidOpenTextDocumentParams {
-                text_document: TextDocumentItem {
-                    uri: manifest_uri.clone(),
-                    language_id: "toml".to_string(),
-                    version: 1,
-                    text: manifest_text.to_string(),
-                },
-            },
-        )))
-        .unwrap();
-
-    let request_id = lsp_server::RequestId::from(1);
-    client
-        .sender
-        .send(Message::Request(Request::new(
-            request_id.clone(),
-            DocumentDiagnosticRequest::METHOD.to_string(),
-            DocumentDiagnosticParams {
-                text_document: TextDocumentIdentifier { uri: manifest_uri },
-                identifier: None,
-                previous_result_id: None,
-                work_done_progress_params: WorkDoneProgressParams::default(),
-                partial_result_params: Default::default(),
-            },
-        )))
-        .unwrap();
-
-    let (_, diagnostics) = recv_document_diagnostics(&client, request_id);
-    assert!(
-        diagnostics.iter().any(|diag| diag.message.contains("unknown field")),
-        "manifest should receive TOML schema diagnostics: {diagnostics:?}"
-    );
-
-    shutdown_test_server(&client, server_thread);
-}
-
-#[test]
-fn project_manifest_reports_missing_path_diagnostics() {
-    let pull_caps = ClientCapabilities {
-        text_document: Some(TextDocumentClientCapabilities {
-            diagnostic: Some(DiagnosticClientCapabilities::default()),
-            ..Default::default()
-        }),
-        ..Default::default()
-    };
-    let temp_dir = TempDir::new("manifest-missing-path-diagnostics");
-    let manifest_text = "sources = [\"missing\"]\nexclude = [\"future-generated\"]\n";
-    let manifest_path = temp_dir.path().join("vizsla.toml");
-    fs::write(&manifest_path, manifest_text).unwrap();
-
-    let root_path = temp_dir.path().to_path_buf();
-    let opt = Opt {
-        process_name: "vizsla-test".to_string(),
-        log: "error".to_string(),
-        log_filename: None,
-    };
-    let config = config::Config::new(
-        opt,
-        root_path.clone(),
-        pull_caps,
-        vec![root_path],
-        UserConfig::default(),
-        Vec::new(),
-    );
-
-    let (server, client) = Connection::memory();
-    let server_thread = thread::spawn(move || main_loop::main_loop(config, server));
-    let manifest_uri = to_proto::url_from_abs_path(manifest_path.as_path()).unwrap();
-
-    client
-        .sender
-        .send(Message::Notification(Notification::new(
-            DidOpenTextDocument::METHOD.to_string(),
-            DidOpenTextDocumentParams {
-                text_document: TextDocumentItem {
-                    uri: manifest_uri.clone(),
-                    language_id: "toml".to_string(),
-                    version: 1,
-                    text: manifest_text.to_string(),
-                },
-            },
-        )))
-        .unwrap();
-
-    let request_id = lsp_server::RequestId::from(1);
-    client
-        .sender
-        .send(Message::Request(Request::new(
-            request_id.clone(),
-            DocumentDiagnosticRequest::METHOD.to_string(),
-            DocumentDiagnosticParams {
-                text_document: TextDocumentIdentifier { uri: manifest_uri },
-                identifier: None,
-                previous_result_id: None,
-                work_done_progress_params: WorkDoneProgressParams::default(),
-                partial_result_params: Default::default(),
-            },
-        )))
-        .unwrap();
-
-    let (_, diagnostics) = recv_document_diagnostics(&client, request_id);
-    assert_eq!(
-        diagnostics.iter().filter(|diag| diag.message.contains("path does not exist")).count(),
-        1,
-        "only non-exclude missing paths should be reported: {diagnostics:?}"
-    );
-    let diagnostic = diagnostics
-        .iter()
-        .find(|diag| diag.message.contains("manifest path does not exist for `sources`"))
-        .expect("sources missing path diagnostic expected");
-    assert_eq!(diagnostic.severity, Some(lsp_types::DiagnosticSeverity::WARNING));
-    assert_eq!(diagnostic.range.start, Position { line: 0, character: 12 });
-    assert_eq!(diagnostic.range.end, Position { line: 0, character: 19 });
-
-    shutdown_test_server(&client, server_thread);
-}
-
-#[test]
-fn project_manifest_completes_top_level_fields() {
-    let temp_dir = TempDir::new("manifest-completion");
-    let manifest_text = "sou";
-    let manifest_path = temp_dir.path().join("vizsla.toml");
-    fs::write(&manifest_path, manifest_text).unwrap();
-
-    let root_path = temp_dir.path().to_path_buf();
-    let opt = Opt {
-        process_name: "vizsla-test".to_string(),
-        log: "error".to_string(),
-        log_filename: None,
-    };
-    let config = config::Config::new(
-        opt,
-        root_path.clone(),
-        ClientCapabilities::default(),
-        vec![root_path],
-        UserConfig::default(),
-        Vec::new(),
-    );
-
-    let (server, client) = Connection::memory();
-    let server_thread = thread::spawn(move || main_loop::main_loop(config, server));
-    let manifest_uri = to_proto::url_from_abs_path(manifest_path.as_path()).unwrap();
-
-    client
-        .sender
-        .send(Message::Notification(Notification::new(
-            DidOpenTextDocument::METHOD.to_string(),
-            DidOpenTextDocumentParams {
-                text_document: TextDocumentItem {
-                    uri: manifest_uri.clone(),
-                    language_id: "toml".to_string(),
-                    version: 1,
-                    text: manifest_text.to_string(),
-                },
-            },
-        )))
-        .unwrap();
-
-    let request_id = lsp_server::RequestId::from(1);
-    client
-        .sender
-        .send(Message::Request(Request::new(
-            request_id.clone(),
-            Completion::METHOD.to_string(),
-            lsp_types::CompletionParams {
-                text_document_position: TextDocumentPositionParams {
-                    text_document: TextDocumentIdentifier { uri: manifest_uri },
-                    position: Position { line: 0, character: 3 },
-                },
-                work_done_progress_params: WorkDoneProgressParams::default(),
-                partial_result_params: Default::default(),
-                context: None,
-            },
-        )))
-        .unwrap();
-
-    let completions: Option<lsp_types::CompletionResponse> =
-        recv_response(&client, request_id, "completion");
-    let items = match completions {
-        Some(lsp_types::CompletionResponse::Array(items)) => items,
-        other => panic!("expected completion array, got {other:?}"),
-    };
-    let sources = items
-        .into_iter()
-        .find(|item| item.label == "sources")
-        .expect("sources completion should be present");
-    let edit = match sources.text_edit {
-        Some(lsp_types::CompletionTextEdit::Edit(edit)) => edit,
-        other => panic!("expected text edit, got {other:?}"),
-    };
-    assert_eq!(edit.new_text, "sources = [\"rtl\"]");
-    assert_eq!(edit.range.start, Position { line: 0, character: 0 });
-    assert_eq!(edit.range.end, Position { line: 0, character: 3 });
-
-    shutdown_test_server(&client, server_thread);
-}
-
-#[test]
-fn project_manifest_completes_only_missing_top_level_fields() {
-    let temp_dir = TempDir::new("manifest-missing-field-completion");
-    let manifest_text = "sources = []\ninc";
-    let manifest_path = temp_dir.path().join("vizsla.toml");
-    fs::write(&manifest_path, manifest_text).unwrap();
-
-    let root_path = temp_dir.path().to_path_buf();
-    let opt = Opt {
-        process_name: "vizsla-test".to_string(),
-        log: "error".to_string(),
-        log_filename: None,
-    };
-    let config = config::Config::new(
-        opt,
-        root_path.clone(),
-        ClientCapabilities::default(),
-        vec![root_path],
-        UserConfig::default(),
-        Vec::new(),
-    );
-
-    let (server, client) = Connection::memory();
-    let server_thread = thread::spawn(move || main_loop::main_loop(config, server));
-    let manifest_uri = to_proto::url_from_abs_path(manifest_path.as_path()).unwrap();
-
-    client
-        .sender
-        .send(Message::Notification(Notification::new(
-            DidOpenTextDocument::METHOD.to_string(),
-            DidOpenTextDocumentParams {
-                text_document: TextDocumentItem {
-                    uri: manifest_uri.clone(),
-                    language_id: "toml".to_string(),
-                    version: 1,
-                    text: manifest_text.to_string(),
-                },
-            },
-        )))
-        .unwrap();
-
-    let request_id = lsp_server::RequestId::from(1);
-    client
-        .sender
-        .send(Message::Request(Request::new(
-            request_id.clone(),
-            Completion::METHOD.to_string(),
-            lsp_types::CompletionParams {
-                text_document_position: TextDocumentPositionParams {
-                    text_document: TextDocumentIdentifier { uri: manifest_uri },
-                    position: Position { line: 1, character: 3 },
-                },
-                work_done_progress_params: WorkDoneProgressParams::default(),
-                partial_result_params: Default::default(),
-                context: None,
-            },
-        )))
-        .unwrap();
-
-    let completions: Option<lsp_types::CompletionResponse> =
-        recv_response(&client, request_id, "completion");
-    let items = match completions {
-        Some(lsp_types::CompletionResponse::Array(items)) => items,
-        other => panic!("expected completion array, got {other:?}"),
-    };
-    let labels = items.iter().map(|item| item.label.as_str()).collect::<Vec<_>>();
-
-    assert!(!labels.contains(&"sources"));
-    assert!(labels.contains(&"include_dirs"));
-
-    shutdown_test_server(&client, server_thread);
-}
-
-#[test]
-fn project_manifest_completes_path_values() {
-    let temp_dir = TempDir::new("manifest-path-completion");
-    fs::create_dir_all(temp_dir.path().join("rtl")).unwrap();
-    fs::write(temp_dir.path().join("top.sv"), "module top; endmodule\n").unwrap();
-    let manifest_text = "sources = [\"r\"]\n";
-    let manifest_path = temp_dir.path().join("vizsla.toml");
-    fs::write(&manifest_path, manifest_text).unwrap();
-
-    let root_path = temp_dir.path().to_path_buf();
-    let opt = Opt {
-        process_name: "vizsla-test".to_string(),
-        log: "error".to_string(),
-        log_filename: None,
-    };
-    let config = config::Config::new(
-        opt,
-        root_path.clone(),
-        ClientCapabilities::default(),
-        vec![root_path],
-        UserConfig::default(),
-        Vec::new(),
-    );
-
-    let (server, client) = Connection::memory();
-    let server_thread = thread::spawn(move || main_loop::main_loop(config, server));
-    let manifest_uri = to_proto::url_from_abs_path(manifest_path.as_path()).unwrap();
-
-    client
-        .sender
-        .send(Message::Notification(Notification::new(
-            DidOpenTextDocument::METHOD.to_string(),
-            DidOpenTextDocumentParams {
-                text_document: TextDocumentItem {
-                    uri: manifest_uri.clone(),
-                    language_id: "toml".to_string(),
-                    version: 1,
-                    text: manifest_text.to_string(),
-                },
-            },
-        )))
-        .unwrap();
-
-    let request_id = lsp_server::RequestId::from(1);
-    client
-        .sender
-        .send(Message::Request(Request::new(
-            request_id.clone(),
-            Completion::METHOD.to_string(),
-            lsp_types::CompletionParams {
-                text_document_position: TextDocumentPositionParams {
-                    text_document: TextDocumentIdentifier { uri: manifest_uri },
-                    position: Position { line: 0, character: 13 },
-                },
-                work_done_progress_params: WorkDoneProgressParams::default(),
-                partial_result_params: Default::default(),
-                context: None,
-            },
-        )))
-        .unwrap();
-
-    let completions: Option<lsp_types::CompletionResponse> =
-        recv_response(&client, request_id, "path completion");
-    let items = match completions {
-        Some(lsp_types::CompletionResponse::Array(items)) => items,
-        other => panic!("expected completion array, got {other:?}"),
-    };
-    let rtl = items
-        .into_iter()
-        .find(|item| item.label == "rtl/")
-        .expect("rtl directory completion should be present");
-    let edit = match rtl.text_edit {
-        Some(lsp_types::CompletionTextEdit::Edit(edit)) => edit,
-        other => panic!("expected text edit, got {other:?}"),
-    };
-    assert_eq!(edit.new_text, "rtl/");
-    assert_eq!(edit.range.start, Position { line: 0, character: 12 });
-    assert_eq!(edit.range.end, Position { line: 0, character: 13 });
-
-    shutdown_test_server(&client, server_thread);
-}
-
-#[test]
-fn project_manifest_goes_to_path_values() {
-    let temp_dir = TempDir::new("manifest-path-definition");
-    let rtl_dir = temp_dir.path().join("rtl");
-    fs::create_dir_all(&rtl_dir).unwrap();
-    let target_path = rtl_dir.join("top.sv");
-    fs::write(&target_path, "module top; endmodule\n").unwrap();
-    let manifest_text = "sources = [\"rtl/top.sv\"]\n";
-    let manifest_path = temp_dir.path().join("vizsla.toml");
-    fs::write(&manifest_path, manifest_text).unwrap();
-
-    let root_path = temp_dir.path().to_path_buf();
-    let opt = Opt {
-        process_name: "vizsla-test".to_string(),
-        log: "error".to_string(),
-        log_filename: None,
-    };
-    let config = config::Config::new(
-        opt,
-        root_path.clone(),
-        ClientCapabilities::default(),
-        vec![root_path],
-        UserConfig::default(),
-        Vec::new(),
-    );
-
-    let (server, client) = Connection::memory();
-    let server_thread = thread::spawn(move || main_loop::main_loop(config, server));
-    let manifest_uri = to_proto::url_from_abs_path(manifest_path.as_path()).unwrap();
-    let target_uri = to_proto::url_from_abs_path(target_path.as_path()).unwrap();
-
-    client
-        .sender
-        .send(Message::Notification(Notification::new(
-            DidOpenTextDocument::METHOD.to_string(),
-            DidOpenTextDocumentParams {
-                text_document: TextDocumentItem {
-                    uri: manifest_uri.clone(),
-                    language_id: "toml".to_string(),
-                    version: 1,
-                    text: manifest_text.to_string(),
-                },
-            },
-        )))
-        .unwrap();
-
-    let request_id = lsp_server::RequestId::from(1);
-    client
-        .sender
-        .send(Message::Request(Request::new(
-            request_id.clone(),
-            GotoDefinition::METHOD.to_string(),
-            GotoDefinitionParams {
-                text_document_position_params: TextDocumentPositionParams {
-                    text_document: TextDocumentIdentifier { uri: manifest_uri },
-                    position: Position { line: 0, character: 16 },
-                },
-                work_done_progress_params: WorkDoneProgressParams::default(),
-                partial_result_params: Default::default(),
-            },
-        )))
-        .unwrap();
-
-    let definition: Option<GotoDefinitionResponse> =
-        recv_response(&client, request_id, "manifest path definition");
-    let definition = definition.expect("path definition should resolve");
-    let location = match definition {
-        GotoDefinitionResponse::Scalar(location) => location,
-        other => panic!("expected scalar location, got {other:?}"),
-    };
-    assert_eq!(location.uri, target_uri);
-    assert_eq!(location.range.start, Position { line: 0, character: 0 });
-
-    shutdown_test_server(&client, server_thread);
-}
-
-#[test]
-fn project_manifest_links_path_values() {
-    let temp_dir = TempDir::new("manifest-path-links");
-    let rtl_dir = temp_dir.path().join("rtl");
-    fs::create_dir_all(&rtl_dir).unwrap();
-    let manifest_text = "sources = [\"rtl\"]\n";
-    let manifest_path = temp_dir.path().join("vizsla.toml");
-    fs::write(&manifest_path, manifest_text).unwrap();
-
-    let root_path = temp_dir.path().to_path_buf();
-    let opt = Opt {
-        process_name: "vizsla-test".to_string(),
-        log: "error".to_string(),
-        log_filename: None,
-    };
-    let config = config::Config::new(
-        opt,
-        root_path.clone(),
-        ClientCapabilities::default(),
-        vec![root_path],
-        UserConfig::default(),
-        Vec::new(),
-    );
-
-    let (server, client) = Connection::memory();
-    let server_thread = thread::spawn(move || main_loop::main_loop(config, server));
-    let manifest_uri = to_proto::url_from_abs_path(manifest_path.as_path()).unwrap();
-    let target_uri = to_proto::url_from_abs_path(rtl_dir.as_path()).unwrap();
-
-    client
-        .sender
-        .send(Message::Notification(Notification::new(
-            DidOpenTextDocument::METHOD.to_string(),
-            DidOpenTextDocumentParams {
-                text_document: TextDocumentItem {
-                    uri: manifest_uri.clone(),
-                    language_id: "toml".to_string(),
-                    version: 1,
-                    text: manifest_text.to_string(),
-                },
-            },
-        )))
-        .unwrap();
-
-    let request_id = lsp_server::RequestId::from(1);
-    client
-        .sender
-        .send(Message::Request(Request::new(
-            request_id.clone(),
-            DocumentLinkRequest::METHOD.to_string(),
-            DocumentLinkParams {
-                text_document: TextDocumentIdentifier { uri: manifest_uri },
-                work_done_progress_params: WorkDoneProgressParams::default(),
-                partial_result_params: Default::default(),
-            },
-        )))
-        .unwrap();
-
-    let links: Option<Vec<lsp_types::DocumentLink>> =
-        recv_response(&client, request_id, "manifest path links");
-    let links = links.expect("document links should return a response");
-    assert_eq!(links.len(), 1);
-    assert_eq!(links[0].target.as_ref(), Some(&target_uri));
-    assert_eq!(links[0].range.start, Position { line: 0, character: 12 });
-    assert_eq!(links[0].range.end, Position { line: 0, character: 15 });
-
-    shutdown_test_server(&client, server_thread);
-}
-
-#[test]
-fn project_manifest_hovers_top_level_fields() {
-    let temp_dir = TempDir::new("manifest-hover");
-    let manifest_text = "sources = [\"rtl\"]\n";
-    let manifest_path = temp_dir.path().join("vizsla.toml");
-    fs::write(&manifest_path, manifest_text).unwrap();
-
-    let root_path = temp_dir.path().to_path_buf();
-    let opt = Opt {
-        process_name: "vizsla-test".to_string(),
-        log: "error".to_string(),
-        log_filename: None,
-    };
-    let config = config::Config::new(
-        opt,
-        root_path.clone(),
-        ClientCapabilities::default(),
-        vec![root_path],
-        UserConfig::default(),
-        Vec::new(),
-    );
-
-    let (server, client) = Connection::memory();
-    let server_thread = thread::spawn(move || main_loop::main_loop(config, server));
-    let manifest_uri = to_proto::url_from_abs_path(manifest_path.as_path()).unwrap();
-
-    client
-        .sender
-        .send(Message::Notification(Notification::new(
-            DidOpenTextDocument::METHOD.to_string(),
-            DidOpenTextDocumentParams {
-                text_document: TextDocumentItem {
-                    uri: manifest_uri.clone(),
-                    language_id: "toml".to_string(),
-                    version: 1,
-                    text: manifest_text.to_string(),
-                },
-            },
-        )))
-        .unwrap();
-
-    let request_id = lsp_server::RequestId::from(1);
-    client
-        .sender
-        .send(Message::Request(Request::new(
-            request_id.clone(),
-            HoverRequest::METHOD.to_string(),
-            HoverParams {
-                text_document_position_params: TextDocumentPositionParams {
-                    text_document: TextDocumentIdentifier { uri: manifest_uri },
-                    position: Position { line: 0, character: 1 },
-                },
-                work_done_progress_params: WorkDoneProgressParams::default(),
-            },
-        )))
-        .unwrap();
-
-    let hover: Option<Hover> = recv_response(&client, request_id, "manifest hover");
-    let hover = hover.expect("sources field should have hover");
-    let text = match hover.contents {
-        lsp_types::HoverContents::Markup(markup) => markup.value,
-        other => panic!("expected markup hover, got {other:?}"),
-    };
-
-    assert!(text.contains("Source scan roots"), "unexpected hover text: {text}");
-    assert_eq!(hover.range.unwrap().start, Position { line: 0, character: 0 });
-    assert_eq!(hover.range.unwrap().end, Position { line: 0, character: 7 });
-
-    shutdown_test_server(&client, server_thread);
-}
-
-#[test]
-fn project_manifest_returns_document_symbols() {
-    let temp_dir = TempDir::new("manifest-document-symbols");
-    let manifest_text = "sources = [\"rtl\"]\ninclude_dirs = [\"include\"]\n";
-    let manifest_path = temp_dir.path().join("vizsla.toml");
-    fs::write(&manifest_path, manifest_text).unwrap();
-
-    let root_path = temp_dir.path().to_path_buf();
-    let opt = Opt {
-        process_name: "vizsla-test".to_string(),
-        log: "error".to_string(),
-        log_filename: None,
-    };
-    let config = config::Config::new(
-        opt,
-        root_path.clone(),
-        ClientCapabilities::default(),
-        vec![root_path],
-        UserConfig::default(),
-        Vec::new(),
-    );
-
-    let (server, client) = Connection::memory();
-    let server_thread = thread::spawn(move || main_loop::main_loop(config, server));
-    let manifest_uri = to_proto::url_from_abs_path(manifest_path.as_path()).unwrap();
-
-    client
-        .sender
-        .send(Message::Notification(Notification::new(
-            DidOpenTextDocument::METHOD.to_string(),
-            DidOpenTextDocumentParams {
-                text_document: TextDocumentItem {
-                    uri: manifest_uri.clone(),
-                    language_id: "toml".to_string(),
-                    version: 1,
-                    text: manifest_text.to_string(),
-                },
-            },
-        )))
-        .unwrap();
-
-    let request_id = lsp_server::RequestId::from(1);
-    client
-        .sender
-        .send(Message::Request(Request::new(
-            request_id.clone(),
-            DocumentSymbolRequest::METHOD.to_string(),
-            DocumentSymbolParams {
-                text_document: TextDocumentIdentifier { uri: manifest_uri },
-                work_done_progress_params: WorkDoneProgressParams::default(),
-                partial_result_params: Default::default(),
-            },
-        )))
-        .unwrap();
-
-    let symbols: Option<DocumentSymbolResponse> =
-        recv_response(&client, request_id, "manifest document symbols");
-    let symbols = match symbols {
-        Some(DocumentSymbolResponse::Flat(symbols)) => symbols,
-        other => panic!("expected flat document symbols, got {other:?}"),
-    };
-    let names = symbols.iter().map(|symbol| symbol.name.as_str()).collect::<Vec<_>>();
-    assert_eq!(names, ["sources", "include_dirs"]);
-    assert_eq!(symbols[0].kind, lsp_types::SymbolKind::PROPERTY);
-    assert_eq!(symbols[0].location.range.start, Position { line: 0, character: 0 });
-    assert_eq!(symbols[0].location.range.end, Position { line: 0, character: 17 });
 
     shutdown_test_server(&client, server_thread);
 }
